Exit with error code on file open/read failures
Gir_*.toml files in gstreamer were recently renamed to <crate>/Gir.toml
but the CI was still overriding their name with `--gir-file Gir_*.toml`.
This leads to "no such file or directory" errors which were promptly
printed to the console and unchecked crates as a result, but the CI job
itself did not receive a non-zero exit code and happily obliged.

Instead of .expect'ing these "common" cases, thread the error back up
through Result<()> to have a readable representation instead of a panic
and eventual stacktrace to wade through.